feat(web): move Jingles / SFX / Beds to their own Imaging admin page
Jingles, Sound FX, and Beds were three sections inside /admin/settings, but
they're asset management (create / upload / delete audio), not configuration —
the same category as Library, Shows, Personas, and Skills, all of which are
top-level nav items. Give them their own home.

- New /admin/imaging page (ImagingPanel) hosting the three as tabs
  (Jingles · SFX · Beds), following the /admin/connect tab pattern
  (Seg control + ?tab= deep-link). Tab labels carry live counts.
- Move JinglesSection / SfxSection / BedsSection into components/admin/imaging/,
  plus their asset types into imaging/types.ts. ImagingPanel absorbs the state
  and handlers these sections needed (previously in SettingsPanel).
- Slim JinglesSection off the whole FormState — it only ever touched
  jingleRatio, so it now takes that one string + a setter, sparing the new page
  Settings' ~200-line form hydration.
- SettingsPanel: drop the three sections, their state/handlers/dialogs, and
  jingleRatio from FormState (~240 lines lighter). Old
  /admin/settings?section={jingles,sfx,beds} bookmarks redirect to the matching
  Imaging tab.
- Add Imaging (Podcast icon) to the Programming nav group.

Pure front-end reorganisation — controller /jingles, /sfx, /beds routes
unchanged.
