+++ web/components/admin/AdminShell.tsx
@@ -23,6 +23,7 @@ import {
   Plug,
   Coffee,
   MessageCircle,
+  Podcast,
 } from 'lucide-react';
 import { useAdminAuth } from '../../lib/adminAuth';
 import type { SignInResult } from '../../lib/adminAuth';
@@ -69,6 +70,7 @@ const NAV_SECTIONS: NavSection[] = [
       { href: '/admin/shows', id: 'shows', label: 'Shows', icon: CalendarClock },
       { href: '/admin/personas', id: 'personas', label: 'Personas', icon: Drama },
       { href: '/admin/skills', id: 'skills', label: 'Skills', icon: Sparkles },
+      { href: '/admin/imaging', id: 'imaging', label: 'Imaging', icon: Podcast },
     ],
   },
   {
